Show group detection conditions first
Currently, when outputting from group detection, the name of the
buildpack is displayed ahead of the condition. For example:

  Buildpack-1: pass
  Buildpack-with-longer-name: pass
  BP-short: fail

As the example shows, conditions are not vertically aligned. This
makes it harder to visually scan for 'pass', 'fail' etc. It also
makes it harder to write scripts that automatically parse lifecycle
output.

This commit puts the conditions ahead of the buildpack name:

  pass: Buildpack-1
  pass: Buildpack-with-longer-name
  fail: BP-short

Signed-off-by: Jacques Chester <jchester@pivotal.io>