Concise Output
Previously the output of the detect phase was way too verbose printing the id
and version of every buildpack and whether it passed or failed.  Given large
buildpack groups this was a large block of text that was difficult to visually
parse for useful information.

A recent change to the output then simplified it to describe only the number
of buildpacks that were participating in the build phase, but removed their id
and version.  This version of the output ended up being too terse, removing
the genuinely useful information of what buildpacks were manipulating your
application.

This change splits the difference between the two previous attempts.  It
prints the number of buildpacks participating out of the number of options
(implying that the others failed or skipped) and then prints the id and
version of each of the participating buildpacks.  The detailed information
about the non-participating buildpacks is still useful in some circumstances,
and should be added back once logging levels are implemented in the lifecycle.
A recommendation would be the current output at Info, and full output at
Debug.

Signed-off-by: Ben Hale <bhale@pivotal.io>