Fix scopes on auth when app/run images are on different registries
This bumps the github.com/google/go-containerregistry dependency so
that a fix for supporting app and run images on from different
registries is pulled in.

In particular, https://github.com/google/go-containerregistry/pull/604
is the fix. Previously, when the app and run image are on different
registries, the app image registry would be asked for the pull scope
of the run image which will fail since it's not present.

Signed-off-by: Nan Zhong <nan@notanumber.io>