Add restorer acceptance tests
Following the existing acceptance test patterns, the new acceptance test will use the test data mounted into an image to prove restorer is properly restoring data as expected.

The contents of the test data was created in-container and copied out of the container using the buildpacks in the testdata folders.

Signed-off-by: Jesse Brown <jabrown85@gmail.com>