Build lifecycle binaries and images for linux/arm64
This changes make build to build for linux/amd64, windows/amd64, and
newly, linux/arm64. Binaries are built to paths like
out/$GOOS-$GOARCH/lifecycle, etc.

This also changes the CI image build process to build and attach the
amd64 binary to the build, and to include the linux/arm64 image in the
manifest list.

Acceptance tests were updated to build arch-specific binaries and to
look for them in the new correct place.

Signed-off-by: Jason Hall <jasonhall@redhat.com>
Signed-off-by: Juan Bustamante <jbustamante@vmware.com>