Try to fix pack acceptance, continued
Use a released version of pack when running pack acceptance (avoids v0.28.0-rc1).

In lifecycle CI, pack acceptance tests are failing on Windows when the pack version is v0.28.0-rc1.
The previous commit, which attempted to make lifecycle CI match pack CI by using go 1.17,
  did not fix the issue.
It was manually confirmed on a Windows dev worker that pack acceptance passes
  with the same pack version and the same lifecycle version that are failing in CI.
Therefore, the failling tests do not appear to relate to the lifecycle,
  and are instead likely due to other unknown differences in the CI environments that should be determined later.

Signed-off-by: Natalie Arellano <narellano@vmware.com>