Fix draft release workflow by giving GITHUB_TOKEN additional permissions
In https://github.com/buildpacks/lifecycle/pull/1201 we explicitly granted
the needed permissions to the token, and we updated the repo settings to be the most restrictive.
Now that the repo has been updated, the draft-release workflow is broken
for minor versions less than 0.18.x.

This change will allow us to continue to patch 0.17.x.

Signed-off-by: Natalie Arellano <narellano@vmware.com>